게시글과 댓글의 수정/삭제 권한 부여하기

+++ app/controllers/foods_controller.rb
@@ -50,6 +50,10 @@ class FoodsController < ApplicationController
 
   def edit
      @post = Post.find(params[:id])
+     if @post.user_id != session[:user_id]
+      flash[:alert] = "수정권한이 없습니다."
+      redirect_to :back
+    end
   end
 
   
@@ -69,11 +73,17 @@ def edit_complete
   end
 
  def delete_complete
-    post = Post.find(params[:id])
-    post.destroy
-    flash[:alert] = "삭제되었습니다."
-    redirect_to "/"
+     post = Post.find(params[:id])
+    if post.user_id == session[:user_id]
+      post.destroy
+      flash[:alert] = "삭제되었습니다"
+      redirect_to "/"
+    else
+      flash[:alert] = "삭제권한이 없습니다"
+      redirect_to :back
+    end
   end
+
 
    before_action :login_check
   skip_before_action  :login_check, :only => [:posts, :posts_category, :show]
@@ -94,9 +104,16 @@ def edit_complete
 
    def delete_comment_complete
     comment = Comment.find(params[:id])
-    comment.destroy
-    flash[:alert] = "댓글이 삭제되었습니다."
-    redirect_to "/foods/show/#{comment.post_id}"
+    if post.user_id == session[:user_id]
+      post.destroy
+      flash[:alert] = "댓글이 삭제되었습니다"
+      redirect_to "/foods/show/#{comment.post_id}}"
+    else
+      flash[:alert] = "해당 댓글의 삭제권한이 없습니다"
+      redirect_to :back
+    end
   end
   
+
+
 end
